+++ app/code/Docker/LogStdout/etc/di.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <preference for="Psr\Log\LoggerInterface" type="Docker\LogStdout\Logger\Monolog" />
+    <type name="Docker\LogStdout\Logger\Handler\Base">
+        <arguments>
+            <argument name="filesystem" xsi:type="object">Magento\Framework\Filesystem\Driver\File</argument>
+        </arguments>
+    </type>
+    <type name="Docker\LogStdout\Logger\Handler\System">
+        <arguments>
+            <argument name="filesystem" xsi:type="object">Magento\Framework\Filesystem\Driver\File</argument>
+        </arguments>
+    </type>
+    <type name="Docker\LogStdout\Logger\Monolog">
+        <arguments>
+            <argument name="name" xsi:type="string">main</argument>
+            <argument name="handlers"  xsi:type="array">
+                <item name="system" xsi:type="object">Docker\LogStdout\Logger\Handler\System</item>
+                <item name="debug" xsi:type="object">Docker\LogStdout\Logger\Handler\Debug</item>
+                <item name="syslog" xsi:type="object">Docker\LogStdout\Logger\Handler\Syslog</item>
+            </argument>
+        </arguments>
+    </type>
+    <type name="Docker\LogStdout\Logger\Handler\Syslog">
+        <arguments>
+            <argument name="ident" xsi:type="string">Magento</argument>
+        </arguments>
+    </type>
+</config>
