+++ app/code/Docker/LogStdout/registration.php
@@ -1,6 +1,6 @@
 <?php
 \Magento\Framework\Component\ComponentRegistrar::register(
     \Magento\Framework\Component\ComponentRegistrar::MODULE,
-    'My_Log',
+    'Docker_LogStdout',
     __DIR__
 );
