--- app/code/My/Log/Logger/Handler.php
@@ -1,17 +1,0 @@
-<?php
-namespace My\Log\Logger;
-
-class Handler extends \Magento\Framework\Logger\Handler\Base
-{
-    /**
-     * Logging level
-     * @var int
-     */
-    protected $loggerType = Logger::INFO;
-
-    /**
-     * File name
-     * @var string
-     */
-    protected $fileName = '/var/log/my_log.log';
-}
