--- app/code/My/Log/Logger/Logger.php
@@ -1,6 +1,0 @@
-<?php
-namespace My\Log\Logger;
-
-class Logger extends \Monolog\Logger
-{
-}
