--- app/code/My/Log/README.md
@@ -1,2 +1,0 @@
-- [Magento2 Custom Log File Using Monolog - Webkul Blog](https://webkul.com/blog/magento2-custom-log-file-using-monolog/)
-- [webkul/magento2_customlog: Magento2 Custom Log File Using Monolog](https://github.com/webkul/magento2_customlog)
