--- app/code/My/Log/etc/di.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <type name="My\Log\Logger\Handler">
-        <arguments>
-            <argument name="filesystem" xsi:type="object">Magento\Framework\Filesystem\Driver\File</argument>
-        </arguments>
-    </type>
-    <type name="My\Log\Logger\Logger">
-        <arguments>
-            <argument name="name" xsi:type="string">customLogHandler</argument>
-            <argument name="handlers"  xsi:type="array">
-                <item name="system" xsi:type="object">My\Log\Logger\Handler</item>
-            </argument>
-        </arguments>
-    </type>
-</config>
