--- app/code/My/Log/etc/module.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" ?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="My_Log" setup_version="2.0.0" />
-</config>
