--- app/code/My/LogStdout/etc/di.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <preference for="Magento\Framework\Logger\Handler\Base" type="My\LogStdout\Logger\Handler\Base" />
-</config>
