--- app/code/My/LogStdout/registration.php
@@ -1,6 +1,0 @@
-<?php
-\Magento\Framework\Component\ComponentRegistrar::register(
-    \Magento\Framework\Component\ComponentRegistrar::MODULE,
-    'My_LogStdout',
-    __DIR__
-);
